When a bound socket accepts a peer, remove it from the bindlisten tracker; it is tracked by the bind tracker.
Explicitly free bind port numbers that are no longer in use so that the
BindPortAssigner can use the numbers again.  Right now this makes no
functional difference as the BindPortAssigner is trivial, but a more
sophisticated implementation will use these hooks.

Allow the user to configure the port range that will be used when the
BindPortAssigner hands out ports to bind requests.

Allow the user to disable support for bind requests entirely.