Optimize the non-splice case by trying to send synchronously before trying to send asynchronously to avoid epoll-related syscall overhead.
A caveat is that at least on Linux, we will see a cap on the maximum
filled buffer size that will be less than the maximum buffer size.  Thus,
we must watermark for using splice against a lower value than the maximum
buffer size.  Arbitrarily from observations, a factor of 3/4 is used.

The optimal buffer sizes may need further tuning, but the basic code
works.