Create a pipe_close() helper to ensure that when a pipe is closed that the socket writing to it has its operations cancelled before the close and receives shut down.  Use it in close_client_socket() and close_remote_socket().
terminate_client() and terminate_remote() no longer need to cancel and
shutdown_receive the writing socket because it is already done by
close_client_socket() and close_remote_socket().

In the socket read handlers that use splice, the wrong
flushPipeTo(Remote|Client)() function was being called, causing the wrong
pipe to be drained.  Use the correct function, which also eliminates the
need for using a strand.post().

When splicing a socket to a pipe, if we receive EAGAIN and the pipe has data
in it, flush the pipe before continuing to read from the socket in case the
pipe is full.  If the pipe is empty, then terminate the connection to avoid
infinite busy-polling.