flushPipeTo(Client|Remote)() callers can now be simplified.
terminate_(client|remote)() does not need to post flushPipeTo(Client|Remote)()
because it does not matter in what order the flush runs relative to other
possible jobs on the strand so long as it completes.  It also does not need
to duplicate functionality is now built-in to flushPipeToX().

The EOF case of the X_read_splice() functions can directly call
terminate_(client|remote)() without manually invoking flushPipeToX()
because terminate_X() will perform any necessary flushes anyway.