Use only two timers for periodically kicking out data from the send and receive pipes.  This reduces the number of timerfds that must be tracked and allocated.
Use one strand for all C->R transfers for all SocksClients and use another
strand for all R->C transfers for all SocksClients.  Saves space per
SocksClient and is necessary so that the sets of SocksClients that have
pending data in pipes that needs to be kicked can be updated without
taking explicit synchronization primitives.