flushPipeTo(Client|Remote)() doesn't need to call terminate_(client|remote)() if splicePipeTo(Client|Remote)() fails.  This change was missed in a previous commit.  It doesn't cause any harm to be double-called, but it's unnecessary.
splicePipeTo(Client|Remote)() now has an optional noterm that will prevent
terminate_(client|remote)() from being automatically called on failure.  It
is needed by the lock-batching changes introduced by this commit.

kick(Client|Remote)PipeBG() is introduced.  It will force the data in a
SocksClient pipe to be fully flushed when called.  Unlike
flushPipeTo(Client|Remote)(), it will not close the SocksClient on completion
and it does not assume that it will be called in place of the normal
tcp_(client|remote)_socket_read_splice() function.

kick(Client|Remote)Pipe() will post kick(Client|Remote)PipeBG() if the
pipe is not fully flushed.

kick(Client|Remote)PipeTimer() now takes the conntracker lock once rather
than for each lookup of a SocksClient.  This batching is intended to
reduce contention and bus traffic on the conntracker lock.  More benchmarks
may need to be taken to see whether this is a worthwhile optimization.