Make the connection trackers template classes in preparation for splitting up the SocksClient class to allow for purely static dispatch.
Remove the ct_ member from the connTracker class and create a fetch memberfn
for the ephConnTracker class that removes a specified shared_ptr from the
tracker and returns that shared_ptr to the caller in order to facilitate
move-semantic-based movement of shared_ptrs between trackers.  Use it
instead of the old ct_-based method that toggled shared_ptr refcounts
in the connTracker::store() memberfn.  As a nice side effect, the
sanity check for a non-existent element is no longer an assert() and now
simply terminates the offending Socks session.