Unify the tcp connect and bind trackers.  We can now distinguish via the SocksTCP::is_bind_ member, and the distinction is only useful for find_by_addr() anyway.
Move splice-using TCP connections to a different tracker.  The number of
connections using splice is relatively few.  Even fewer if we switch back to
using read/write when the traffic backs down.  Thus, we can just wake up the
kicktimer when any socket moves to splice-mode, and the timer can just lock
and iterate across the splice tracker, checking the pipe lengths and send
timestamps.

This is a little tricky in practice because we do not take locks for
the timestamp, and we want to avoid deadlocking on the tracker lock if
a connection is terminated in the process of flushing the pipe buffers.
Thus we still need two timers that each run in different strands.