Greatly simplify and improve the pipe buffering flush timeout.
Use forward_list<std::weak_ptr<SocksTCP>>s to track sessions that are
currently using splice.  We keep one list for client->remote and
remote->client directions each.  When we switch to splicing, we commit
a weak_ptr to the appropriate list and kick the matching timer.

When the timer triggers, we go through the list an item at a time.  If
necessary, we flush out the splice buffer.  If the weak_ptr is no longer
valid, or the SockTCP session is no longer splicing that direction, we
remove the weak_ptr from the list.  We continue running the timer until
the list is empty.

This method has absolutely no contention on the conntracker and it
eliminates many temporaries in the previous approach that were necessary
to dance around the conntracker lock.  I can't see any downsides to this
change.

Remove the conntracker_tcp_splice since it's no longer needed, either.  All
of the SocksTCP can live in a single tracker.