Shrink more member variables in SocksInit to reduce memory usage.
Specifically, bitfields can be used for the command type and address type.
The various buffer length-related fields can be shrunk to uint8_t as well.

An extra guard against overflow is placed in the DNS name parsing path.
I am confident that there is no need for the check, but it's best to be
very safe since DNS is rarely used in SOCKS queries.  Therefore, the
check is explicit and does not use assert().