Create the TCP DNS resolver only when needed, and set a timer on creation that is used to destroy the resolver when it has not been used.
Whenever a DNS lookup is queued, increment a sequence counter to indicate
the state transition.  If the DNS resolver timer expires with an unchanging
sequence counter over its lifetime, then the TCP DNS resolver was unused
and can be destroyed to reclaim resources.  Otherwise, the timer is set
again with the current sequence number.