Remove spliceit() function and treat EAGAIN on pipe writes as an error.
We should never encounter EAGAIN when performing socket->pipe writes
since we aggressively drain the pipe before reading more data from
the socket.  EAGAIN in this case would signify a full pipe, which
should never happen and thus should be treated as an error.

In the case of pipe->socket writes, EAGAIN means that the operation
would block.  This not necessarily a fatal error since we do make
an attempt to eagerly flush data out from pipes to the destination
socket before polling for writes on the destination socket.